requirements.txt
Django==4.2.18
dj-database-url==3.0.1
django-allauth==65.10.0
django-storages==1.14.6
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
Pillow==11.3.0
boto3==1.39.7
whitenoise==6.6.0
gunicorn==21.2.0
